tests/wpt/tests/webvtt/parsing/file-parsing/tests/regions-scroll.html

<!doctype html>
<!-- DO NOT EDIT! This file and support/regions-scroll.vtt are generated. -->
<!-- See /webvtt/parsing/file-parsing/README.md -->
<meta charset=utf-8>
<title>WebVTT parser test: regions, scroll</title>
<link rel="help" href="https://w3c.github.io/webvtt/#collect-webvtt-region-settings">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var t = async_test('regions, scroll');
t.step(function(){
    var video = document.createElement('video');
    var track = document.createElement('track');
    assert_true('src' in track, 'track element not supported');
    track.src = 'support/regions-scroll.vtt';
    track['default'] = true;
    track.kind = 'subtitles';
    track.onload = this.step_func(trackLoaded);
    track.onerror = this.step_func(trackError);
    video.appendChild(track);
    document.body.appendChild(video);
});

function trackLoaded(event) {
    var track = event.target;
    var video = track.parentNode;
    var cues = video.textTracks[0].cues;
    {
assert_equals(cues.length, 6);

var regions = Array.from(cues).map(function(cue) {
    return cue.region;
});

var valid_lines = [
    '',
    'up',
    'up',
    '',
    '',
    'up',
].forEach(function(valid, index) {
    assert_equals(regions[index].scroll, valid, 'Failed with region ' + index);
});

    }
    this.done();
}

function trackError(e) {
    assert_unreached('got unexpected error event');
}
</script>
